Refactor
Refactor:

Register Teacher Page
Register Student Page
Update Profile Teacher Page
Update Profile Student Page
Admin -> Drop Student Class Page
