+++ api-academic_information_center/src/students/student.controller.js
@@ -17,8 +17,6 @@ export class StudentController {
     }
   }
 
-  async (){}
-
   async getStudentInfoByAcademicId(req,res,next){
     try{
       const {academicId} = req.user;
