+++ api-academic_information_center/src/students/students.service.js
@@ -20,6 +20,7 @@ export class StudentsService {
           academicId: academicId? academicId:undefined
         },
         select:{
+          id: true,
           academicId: true,
           names: true,
           fatherLastName: true,
