+++ app-students/app/layouts/admin/components/student-subject-item/student-subject-item.css
