+++ app-students/app/layouts/admin/pages/register-student/register-student.html
@@ -1,0 +1,44 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Registro de Estudiante</title>
+  <link rel="stylesheet" href="register-student.css">
+</head>
+<body>
+  <back-button></back-button>
+<div class="page">
+  <div class="form-container">
+    <h1>Registrar Estudiante</h1>
+    <form id="studentForm" autocomplete="on">
+      <div class="form-group">
+        <label for="name">Nombre</label>
+        <input type="text" id="name" name="name" placeholder="Ingresa el nombre" required>
+      </div>
+      <div class="form-group">
+        <label for="fatherLastName">Apellido Paterno</label>
+        <input type="text" id="fatherLastName" name="fatherLastName" placeholder="Ingresa el apellido paterno" required>
+      </div>
+      <div class="form-group">
+        <label for="motherLastName">Apellido Materno</label>
+        <input type="text" id="motherLastName" name="motherLastName" placeholder="Ingresa el apellido materno" required>
+      </div>
+      <button type="submit" id="registerBtn">Registrar</button>
+    </form>
+  </div>
+  <div id="studentInfo" class="student-info" style="display: none;">
+    <h2>Información del Estudiante Registrado</h2>
+    <p><strong>ID:</strong> <span id="academiIdReturned"></span></p>
+    <p><strong>Email:</strong> <span id="emailReturned"></span></p>
+    <p><strong>Contraseña:</strong> <span id="passwordReturned"></span></p>
+    <p><strong>Nombre:</strong> <span id="namesReturned"></span></p>
+    <p><strong>Apellido Paterno:</strong> <span id="fatherLastNameReturned"></span></p>
+    <p><strong>Apellido Materno:</strong> <span id="motherLastNameReturned"></span></p>
+    <button type="submit" id="returToMenuBtn">Regresar al Menú</button>
+</div>
+<div id="responseMessage" class="toast"></div>
+</div>
+<script type="module" src="register-student.js"></script>
+</body>
+</html>
