+++ app-students/app/layouts/admin/pages/teacher-options/teacher-options.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Admin Dashboard</title>
+    <link rel="stylesheet" href="teacher-options.css"> <!-- Agregar tu archivo de estilos -->
+</head>
+<body>
+    <admin-header></admin-header>
+    <main id="admin-main">
+        <div id="options"></div>
+    </main>
+    <!-- footer here -->
+    <script type="module" src="teacher-options.js"></script>
+</body>
+</html>
