+++ app-students/app/layouts/student/pages/options/options.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Student Dashboard</title>
+    <link rel="stylesheet" href="options.css"> <!-- Agregar tu archivo de estilos -->
+</head>
+<body>
+    <student-header></student-header>
+    <main id="student-main">
+        <div id="options"></div>
+    </main>
+    <!-- footer here -->
+    <script type="module" src="options.js"></script>
+</body>
+</html>
