+++ app-students/app/layouts/teacher/pages/options/options.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Teacher Dashboard</title>
+    <link rel="stylesheet" href="options.css"> <!-- Agregar tu archivo de estilos -->
+</head>
+<body>
+    <teacher-header></teacher-header>
+    <main id="teacher-main">
+        <div id="options"></div>
+    </main>
+    <!-- footer here -->
+    <script type="module" src="options.js"></script>
+</body>
+</html>
