+++ app-students/app/shared/router.js
@@ -3,7 +3,12 @@ import { routes } from '../routes.js';
 class Router {
     constructor() {}
     
-    navigate(path) {
+    navigate(path,query={}) {
+        let pathQuery;
+        if (query && Object.keys(query).length > 0) {
+            const queryString = new URLSearchParams(query).toString();
+            pathQuery = `?${queryString}`;
+        }
         const route = routes.find(route => route.path === path);
         if (!route) {
             throw new Error(`Route not found: ${path}`);
@@ -11,7 +16,12 @@ class Router {
         if (!this.#evaluateGuards(route)) {
             return;
         }
-        window.location.pathname = route.page;
+        let fullPath = route.page;
+        if(pathQuery){
+            fullPath = `${fullPath}${pathQuery}`;
+        }
+        const absolutePath = new URL(fullPath, window.location.origin).href;
+        window.location.href = absolutePath;
     }
 
     replace(path) {
@@ -24,7 +34,8 @@ class Router {
                 return;
             }
         });
-        window.location.replace(route.page);
+        const absolutePath = new URL(route.page, window.location.origin).href;
+        window.location.replace(absolutePath);
     }
 
     back() {
